+++ packages/components/components/elvis-accordion/CHANGELOG.md
@@ -1,10 +1,19 @@
 # Elvia Accordion Changelog
 
-## 2.0.4 (23.11.21)
+## 2.0.5 (30.11.21)
+
+### Patch
+
+- Now using elvis-colors package for colors.
+
+## 2.0.4 (30.11.21)
+
+- Web component wrapper fix
 
 ### Bug fix
 
-- Now using elvis-colors package for colors.
+- Improved type safety.
+
 
 ## 2.0.3 (17.11.21)